+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `RoarBot.prototype.commands` property for accessing commands.
 - `RoarBot.prototype.ws` property for accessing the current WebSocket.
+- `RoarBot.prototype.setAccountSettings` method for setting various account options.
 
 ## 1.2.0 - 2024-08-01